Rename to `seedprovider` webhook
